feat: rename Provider to Vendor and refactor some code
