+++ src/abstracts/deployed.ts
@@ -6,20 +6,20 @@
 
 import { OWNER_REQUIRED } from '../errors'
 import { Keyed, Contract, TransactOpts } from '../interfaces'
-import Provider from './provider'
+import Vendor from './vendor'
 
 export default abstract class implements Keyed {
   [key:string]: any
   public abi: any
-  public provider: Provider
+  public vendor: Vendor
   public contract?: Contract
 
   /**
    * @param p - Optional Provider the HOC will use
    * @param a - Compiled abi of the smart contract this HOC wraps
    */
-  protected constructor(p: Provider, a: any) {
-    this.provider = p
+  protected constructor(v: Vendor, a: any) {
+    this.vendor = v
     this.abi = a
   }
 
@@ -30,7 +30,7 @@ export default abstract class implements Keyed {
    * @returns boolean indicating a successful wrapping of the target deployed contract
    */
   protected at(a: string, o?: TransactOpts): boolean {
-    this.contract = this.provider.contract(a, this.abi, o)
+    this.contract = this.vendor.contract(a, this.abi, o)
     return !!this.contract
   }
 }
