+++ src/abstracts/vendor.ts
